requirements.txt
streamlit==1.32.0
opencv-python==3.10.13
numpy==1.24.4
matplotlib==3.8.3
Pillow==10.2.0
